few changes made upto video 29
On the profile page i.e.(profile.php), only the posts that have 0 likes can get deleted.
The posts having more than 0 like, throw an error while deleting them. This is probably beacuse of a foreign key present in post_likes table.

On the index.php page when a comment is added for a particular post, it changes the likes to -1 initially and then the very next comment changes it to +1.

Image does not get posted.